MOV_player/requirements.txt
altgraph==0.17.4
imageio==2.35.1
macholib==1.16.3
modulegraph==0.19.6
numpy==2.1.2
opencv-python==4.10.0.84
pefile==2024.8.26
pillow==10.4.0
py2app==0.28.8
python-vlc==3.0.21203
tifffile==2024.9.20
